+++ lazytelemetry/lazytracing/goroutine_not_lazydev.go
@@ -1,0 +1,7 @@
+//go:build !lazydev
+
+package lazytracing
+
+func currentGoroutineID() uint64 {
+	return 0
+}